Добавление скрытого логгера в прод

+++ Assets/GameResources/Conveyor/Scripts/ConveyorRider.cs
@@ -109,7 +109,7 @@ namespace Features.Conveyor
         {
             if (pathTween != null)
             {
-                if (pathTween.IsActive())
+                if (pathTween.IsActive() && pathTween.active)
                 {
                     pathTween.Kill();
                 }
